+++ studio/commands/__init__.py
@@ -4,6 +4,8 @@ from studio.commands.command import Command
 from studio.commands.command_manager import CommandManager
 from studio.commands.delete_piece_command import DeletePieceCommand
 from studio.commands.duplicate_piece_command import DuplicatePieceCommand
+from studio.commands.edit_board_command import EditBoardCommand
+from studio.commands.edit_piece_command import EditPieceCommand
 from studio.commands.import_boards_command import ImportBoardsCommand
 from studio.commands.import_pieces_command import ImportPiecesCommand
 from studio.commands.move_piece_command import MovePieceCommand
@@ -14,6 +16,8 @@ __all__ = [
     "CommandManager",
     "DeletePieceCommand",
     "DuplicatePieceCommand",
+    "EditBoardCommand",
+    "EditPieceCommand",
     "ImportBoardsCommand",
     "ImportPiecesCommand",
     "MovePieceCommand",